+++ geoextent/__main__.py
@@ -4,8 +4,8 @@ from .lib import extent
 from . import __version__ as current_version
 import logging
 
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
+logging.basicConfig(level=logging.WARNING)
+logger = logging.getLogger("geoextent")
 
 help_description = '''
 geoextent is a Python library for extracting geospatial and temporal extents of a file or a directory of multiple geospatial data formats.
@@ -49,7 +49,7 @@ def get_argparser():
         add_help=False,
         prog='geoextent',
         formatter_class=argparse.RawDescriptionHelpFormatter,
-        usage= "geoextent [-h] [--formats] [--version] [-b] [-t] [-input= '[filepath|input file]']"
+        usage= "geoextent [-h] [--formats] [--version] [--debug] [-b] [-t] [input file]']"
     )
 
     parser.add_argument(
@@ -71,35 +71,40 @@ def get_argparser():
     )
 
     parser.add_argument(
+        '--debug',
+        help='turn on debug logging, alternatively set environment variable GEOEXTENT_DEBUG=1',
+        action='store_true'
+    )
+
+    parser.add_argument(
         '-b', '--bounding-box',
         action='store_true',
+        default=False,
         help='extract spatial extent (bounding box)'
     )
 
     parser.add_argument(
         '-t', '--time-box',
         action='store_true',
+        default=False,
         help='extract temporal extent'
     )
 
     parser.add_argument(
-        '-input=',
+        'files',
         action=readable_file_or_dir,
         default=os.getcwd(),
-        nargs='+',
+        nargs=argparse.REMAINDER,
         help="input file or path"
     )
     
     return parser
 
-def print_help_fun():
+def print_help():
     print(help_description)
     argparser.print_help()
-    print_examples()
-    print_supported_formats()
-
-def print_examples():
     print(help_epilog)
+    print_supported_formats()
 
 def print_supported_formats():
     print(supported_formats)
@@ -113,34 +118,50 @@ argparser = get_argparser()
 def main():
     # Check if there is no arguments, then print help
     if len(sys.argv[1:])==0:
-        print_help_fun()
+        print_help()
         argparser.exit()
 
-    args = vars(argparser.parse_args())
-    logger.debug('Extracting from inputs %s', args['input='])
-
-    if(args['help']):
-        print_help_fun()
-    elif(args['version']):
+    # version, help, and formats must be checked before parse, as otherwise files are required 
+    # but argpase gives an error if allowed to be parsed first
+    if "--help" in sys.argv:
+        print_help()
+        argparser.exit()
+    if "--version" in sys.argv:
         print_version()
-    elif(args['formats']):
+        argparser.exit()
+    if "--formats" in sys.argv:
         print_supported_formats()
-    else:
-        # Check if file is exists happens in parser validation, see readable_file_or_dir
-        if os.path.isfile(os.path.join(os.getcwd(), args['input='])):
-            output = extent.fromFile(args['input='], bbox = args['bounding_box'], tbox = args['time_box'])
-        if os.path.isdir(os.path.join(os.getcwd(), args['input='])):
-            output = extent.fromDirectory(args['input='], bbox = args['bounding_box'], tbox = args['time_box'])
-        
-        if output is None:
-            raise Exception("This file format is not supported")
-
-        # print output
-        if type(output) == list or type(output) == dict:
-            print(str(output))
-        else: 
-            print(output)
+        argparser.exit()
     
+    args = vars(argparser.parse_args())
+    files = args['files']
+    logger.debug('Extracting from inputs %s', files)
+
+    # Set logging level
+    if(args['debug']):
+        logging.getLogger('geoextent').setLevel(logging.DEBUG)
+    if(os.environ.get('GEOEXTENT_DEBUG', None) == "1"):
+        logging.getLogger('geoextent').setLevel(logging.DEBUG)
+
+    # Check if file is exists happens in parser validation, see readable_file_or_dir
+    try:
+        if os.path.isfile(os.path.join(os.getcwd(), files)):
+            output = extent.fromFile(files, bbox = args['bounding_box'], tbox = args['time_box'])
+        if os.path.isdir(os.path.join(os.getcwd(), files)):
+            output = extent.fromDirectory(files, bbox = args['bounding_box'], tbox = args['time_box'])
+    except Exception as e:
+        if(logger.getEffectiveLevel() >= logging.DEBUG):
+            logger.exception(e)
+        sys.exit(1)
+
+    if output is None:
+        raise Exception("Did not find supported files at {}".format(files))
+
+    # print output
+    if type(output) == list or type(output) == dict:
+        print(str(output))
+    else: 
+        print(output)
 
 if __name__ == '__main__':
     main()