+++ tests/test_api_csv.py
@@ -91,11 +91,6 @@ def test_csv_extract_bbox_and_tbox_with_TIME_DATE():
     assert result["bbox"] == [4.3175, 51.434444, 6.574722, 53.217222]
     assert result["tbox"] == ['01.08.2017', '30.09.2019']
 
-def test_csv_extract_bbox_and_tbox_when_both_false():
-    with pytest.raises(Exception) as excinfo:
-        geoextent.fromFile('tests/testdata/csv/cities_NL.csv', bbox=False, tbox=False)
-    assert "Please enter correct arguments" in str(excinfo.value)
-
 def test_csv_extract_bbox_and_tbox_with_cols_diff_order():
     result = geoextent.fromFile('tests/testdata/csv/cities_NL_case1.csv', bbox=True, tbox=True)
     assert "bbox" in result